refactor: extract message edit helpers (normalizeImgGenHtmlForEditing, prepareEditText, restoreEditText) into core/utils
